added SimplePizzaFactory class to remove object creation from PizzaStore class
