+++ src/com/app/creational/FactoryDesign/example/ClamPizza.java
@@ -1,0 +1,4 @@
+package com.app.creational.FactoryDesign.example;
+
+public class ClamPizza extends Pizza {
+}
