+++ src/com/app/creational/FactoryDesign/example/SimplePizzaFactory.java
@@ -1,0 +1,21 @@
+package com.app.creational.FactoryDesign.example;
+
+public class SimplePizzaFactory {
+
+    public Pizza createPizza(String type){
+
+        Pizza pizza = null;
+
+
+        if(type.equals("cheese")) {
+            pizza = new CheesePizza();
+        }else if(type.equals("pepperoni")){
+            pizza = new PepperoniPizza();
+        }else if(type.equals("clam")){
+            pizza = new ClamPizza();
+        }else if(type.equals("veggie")){
+            pizza = new VeggiePizza();
+        }
+        return pizza;
+    }
+}
