+++ src/com/app/creational/FactoryDesign/example/VeggiePizza.java
@@ -1,0 +1,4 @@
+package com.app.creational.FactoryDesign.example;
+
+public class VeggiePizza extends Pizza {
+}
